📝 docs(README.md): add section on using Nodemon for autorestarts
The README.md file was updated to include a new section on using Nodemon for autorestarts. This section provides instructions on how to install Nodemon globally and how to add a script to the "start" command in the package.json file. It also mentions that running "nodemon app.js" directly will not work unless Nodemon is installed globally.